requirements.txt
requests>=2.25.1
pandas>=1.3.0
matplotlib>=3.4.0
python-dotenv>=0.19.0 